pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finqa-bot"
version = "0.1.0"
description = "FinQA chatbot with typed DSL output, hybrid retrieval, two-tier model routing, and vLLM-native inference optimization."
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = { text = "Apache-2.0" }
authors = [{ name = "Rajvi Zala" }]
keywords = ["finqa", "langgraph", "langchain", "vllm", "rag", "numerical-reasoning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.7,<3",
    "pydantic-settings>=2.3,<3",
    "typer>=0.12,<1",
    "rich>=13.7",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "tenacity>=8.5",

    "langchain>=0.3.20,<0.4",
    "langchain-core>=0.3.40,<0.4",
    "langchain-openai>=0.2.10,<0.4",
    "langchain-community>=0.3.15,<0.4",
    "langgraph>=0.2.55,<0.5",
    "langgraph-checkpoint-sqlite>=2.0.1",

    "datasets>=2.20",
    "huggingface-hub>=0.24",
    "transformers>=4.45",
    "tokenizers>=0.20",
    "sentence-transformers>=3.1",
    "rank-bm25>=0.2.2",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26,<3",
    "pandas>=2.2",
    "scipy>=1.13",
    "sympy>=1.13",

    "prometheus-client>=0.20",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.1",
    "gradio>=5.0,<6",

    "tiktoken>=0.7",
    "regex>=2024.7",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
serve = [
    "vllm>=0.6.3",
]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.11",
    "types-pyyaml>=6.0",
    "types-requests>=2.32",
    "hypothesis>=6.112",
]
openai = [
    "openai>=1.40",
]

[project.scripts]
finqa-bot = "finqa_bot.cli:app"

[project.urls]
Homepage = "https://github.com/"
Repository = "https://github.com/"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["finqa_bot*"]

[tool.setuptools.package-data]
"finqa_bot" = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["data", "build", "dist", ".venv"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "PLC", "PLE", "RUF"]
ignore = ["E501", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["N802", "N806", "PLR2004"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
no_implicit_optional = true
check_untyped_defs = true
show_error_codes = true
pretty = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "faiss.*",
    "rank_bm25.*",
    "sentence_transformers.*",
    "gradio.*",
    "sse_starlette.*",
    "vllm.*",
    "prometheus_client.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: long-running tests (deselect with '-m \"not slow\"')",
    "gpu: requires GPU (deselect with '-m \"not gpu\"')",
    "integration: end-to-end smoke tests",
]
